add: Move the API and controller code to /pkg.
