+++ config/crd/bases/intent.security.nimbus.com_securityintents.yaml
@@ -1,0 +1,173 @@
+---
+apiVersion: apiextensions.k8s.io/v1
+kind: CustomResourceDefinition
+metadata:
+  annotations:
+    controller-gen.kubebuilder.io/version: v0.13.0
+  name: securityintents.intent.security.nimbus.com
+spec:
+  group: intent.security.nimbus.com
+  names:
+    kind: SecurityIntent
+    listKind: SecurityIntentList
+    plural: securityintents
+    shortNames:
+    - sit
+    singular: securityintent
+  scope: Namespaced
+  versions:
+  - name: v1
+    schema:
+      openAPIV3Schema:
+        description: SecurityIntent is the Schema for the securityintents API
+        properties:
+          apiVersion:
+            description: 'APIVersion defines the versioned schema of this representation
+              of an object. Servers should convert recognized schemas to the latest
+              internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources'
+            type: string
+          kind:
+            description: 'Kind is a string value representing the REST resource this
+              object represents. Servers may infer this from the endpoint the client
+              submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds'
+            type: string
+          metadata:
+            type: object
+          spec:
+            description: SecurityIntentSpec defines the desired state of SecurityIntent
+            properties:
+              intent:
+                description: Intent defines the security policy details
+                properties:
+                  action:
+                    type: string
+                  description:
+                    type: string
+                  id:
+                    type: string
+                  mode:
+                    type: string
+                  params:
+                    items:
+                      description: Resource defines the resources that the security
+                        policy applies to
+                      properties:
+                        SecurityIntentMatchProtocols:
+                          description: 'Network: MatchProtocols'
+                          items:
+                            description: SecurityIntentMatchProtocol defines a protocol
+                              for network policies
+                            properties:
+                              protocol:
+                                type: string
+                            type: object
+                          type: array
+                        fromCIDRSet:
+                          items:
+                            description: CIDRSet defines CIDR ranges for network policies
+                            properties:
+                              cidr:
+                                type: string
+                            type: object
+                          type: array
+                        matchCapabilities:
+                          description: 'Capabilities: MatchCapabilities'
+                          items:
+                            description: MatchCapability defines a capability for
+                              capabilities policies
+                            properties:
+                              capability:
+                                type: string
+                            type: object
+                          type: array
+                        matchDirectories:
+                          items:
+                            description: MatchDirectory defines a directory for process
+                              or file policies
+                            properties:
+                              dir:
+                                type: string
+                              fromSource:
+                                items:
+                                  properties:
+                                    path:
+                                      type: string
+                                  type: object
+                                type: array
+                            type: object
+                          type: array
+                        matchPaths:
+                          description: 'Process: MatchPaths, MatchDirectories, MatchPatterns
+                            File: MatchPaths, MatchDirectories'
+                          items:
+                            description: MatchPath defines a path for process or file
+                              policies
+                            properties:
+                              path:
+                                type: string
+                            type: object
+                          type: array
+                        matchPatterns:
+                          items:
+                            description: MatchPattern defines a pattern for process
+                              policies
+                            properties:
+                              pattern:
+                                type: string
+                            type: object
+                          type: array
+                        matchSyscalls:
+                          description: 'Syscalls: MatchSyscalls'
+                          items:
+                            description: MatchSyscall defines a syscall for syscall
+                              policies
+                            properties:
+                              syscalls:
+                                items:
+                                  type: string
+                                type: array
+                            type: object
+                          type: array
+                        toPorts:
+                          items:
+                            description: ToPort defines ports and protocols for network
+                              policies
+                            properties:
+                              ports:
+                                items:
+                                  description: Port defines a network port and its
+                                    protocol
+                                  properties:
+                                    port:
+                                      type: string
+                                    protocol:
+                                      type: string
+                                  type: object
+                                type: array
+                            type: object
+                          type: array
+                      type: object
+                    type: array
+                  severity:
+                    type: integer
+                  tags:
+                    items:
+                      type: string
+                    type: array
+                required:
+                - action
+                - id
+                - mode
+                - params
+                type: object
+            required:
+            - intent
+            type: object
+          status:
+            description: SecurityIntentStatus defines the observed state of SecurityIntent
+            type: object
+        type: object
+    served: true
+    storage: true
+    subresources:
+      status: {}
