+++ config/rbac/role.yaml
@@ -1,0 +1,84 @@
+---
+apiVersion: rbac.authorization.k8s.io/v1
+kind: ClusterRole
+metadata:
+  name: manager-role
+rules:
+- apiGroups:
+  - intent.security.nimbus.com
+  resources:
+  - nimbuspolicies
+  verbs:
+  - create
+  - delete
+  - get
+  - list
+  - patch
+  - update
+  - watch
+- apiGroups:
+  - intent.security.nimbus.com
+  resources:
+  - nimbuspolicies/finalizers
+  verbs:
+  - update
+- apiGroups:
+  - intent.security.nimbus.com
+  resources:
+  - nimbuspolicies/status
+  verbs:
+  - get
+  - patch
+  - update
+- apiGroups:
+  - intent.security.nimbus.com
+  resources:
+  - securityintentbindings
+  verbs:
+  - create
+  - delete
+  - get
+  - list
+  - patch
+  - update
+  - watch
+- apiGroups:
+  - intent.security.nimbus.com
+  resources:
+  - securityintentbindings/finalizers
+  verbs:
+  - update
+- apiGroups:
+  - intent.security.nimbus.com
+  resources:
+  - securityintentbindings/status
+  verbs:
+  - get
+  - patch
+  - update
+- apiGroups:
+  - intent.security.nimbus.com
+  resources:
+  - securityintents
+  verbs:
+  - create
+  - delete
+  - get
+  - list
+  - patch
+  - update
+  - watch
+- apiGroups:
+  - intent.security.nimbus.com
+  resources:
+  - securityintents/finalizers
+  verbs:
+  - update
+- apiGroups:
+  - intent.security.nimbus.com
+  resources:
+  - securityintents/status
+  verbs:
+  - get
+  - patch
+  - update
