+++ nimbus-operator/.dockerignore
