+++ nimbus-operator/pkg/receiver/watcher/watch_controller.go
@@ -7,7 +7,7 @@ import (
 	"context"
 	"fmt"
 
-	v1 "github.com/5GSEC/nimbus/nimbus-operator/api/v1"
+	v1 "github.com/5GSEC/nimbus/nimbus-operator/pkg/api/v1"
 	ctrl "sigs.k8s.io/controller-runtime"
 	"sigs.k8s.io/controller-runtime/pkg/client"
 )
